Give each ICMP backend its own file
Adding the Windows path took ping.rs from 289 lines to 422, past the 300
the size guard allows. Splitting by platform is the cut the file was
already asking for: ping.rs keeps the scanner, the result type and the
backend choice, and each ICMP implementation moves next to the platform it
serves.

The "filters replies by identifier and sequence" note moves with the raw
socket code, where it is true. IcmpSendEcho does that matching itself.
